requirements.txt
absl-py==1.4.0
aiobotocore==2.5.0
aiohttp==3.8.4
aioitertools==0.11.0
aiosignal==1.3.1
albumentations==1.3.1
altair==5.0.1
antlr4-python3-runtime==4.9.3
anyio==3.7.0
arrow==1.2.3
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
bleach==6.0.0
blessed==1.20.0
blinker==1.6.2
bokeh==3.1.1
botocore==1.29.76
cachetools==5.3.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.1.0
croniter==1.3.15
cycler==0.11.0
dateutils==0.6.12
decorator==5.1.1
deepdiff==6.3.0
docker==6.1.3
docstring-parser==0.15
einops==0.6.1
exceptiongroup==1.1.1
fastapi==0.98.0
filelock==3.9.0
fonttools==4.40.0
frozenlist==1.3.3
fsspec==2023.4.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.21.0
google-auth-oauthlib==1.0.0
grpcio==1.56.0
h11==0.14.0
huggingface-hub==0.15.1
hydra-core==1.3.2
idna==3.4
imageio==2.31.1
importlib-metadata==6.7.0
importlib-resources==5.12.0
inquirer==3.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.0
jsonargparse==4.22.0
jsonschema==4.17.3
kiwisolver==1.4.4
lazy_loader==0.2
lightning==2.0.4
lightning-api-access==0.0.5
lightning-cloud==0.5.37
lightning-fabric==2.0.4
lightning-utilities==0.8.0
linkify-it-py==2.0.2
lion-pytorch==0.1.2
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdit-py-plugins==0.4.0
mdurl==0.1.2
mpmath==1.2.1
multidict==6.0.4
networkx==3.0rc1
numpy==1.24.1
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python-headless==4.7.0.72
ordered-set==4.1.0
packaging==23.1
pandas==2.0.2
panel==1.1.1
param==1.13.0
Pillow==9.3.0
protobuf==4.23.3
psutil==5.9.5
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.9
pydeck==0.8.1b0
Pygments==2.15.1
PyJWT==2.7.0
Pympler==1.0.1
pyparsing==3.1.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-editor==1.0.4
python-multipart==0.0.6
pytorch-lightning==2.0.4
pytorch-triton==2.1.0+440fd1bf20
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pyviz-comms==2.3.2
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
readchar==4.0.5
redis==4.6.0
regex==2023.6.3
requests==2.28.1
requests-oauthlib==1.3.1
rich==13.4.2
rsa==4.9
s3fs==2023.4.0
safetensors==0.3.1
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.11.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
starlette==0.27.0
starsessions==1.3.0
streamlit==1.24.0
sympy==1.11.1
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorboardX==2.6.1
threadpoolctl==3.1.0
tifffile==2023.4.12
timm==0.9.2
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.1.0.dev20230626+cu121
torchaudio==2.1.0.dev20230626+cu121
torchmetrics==0.11.4
torchvision==0.16.0.dev20230626+cu121
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.30.2
typeshed-client==2.3.0
typing_extensions==4.4.0
tzdata==2023.3
tzlocal==4.3.1
uc-micro-py==1.0.2
urllib3==1.26.13
uvicorn==0.22.0
validators==0.20.0
vit-pytorch==1.2.2
watchdog==3.0.0
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.6.1
websockets==11.0.3
Werkzeug==2.3.6
wrapt==1.15.0
xyzservices==2023.5.0
yarl==1.9.2
zipp==3.15.0
